ecommerce/requirements.txt
black
Django 
djangorestframework  
isort
waitress